pyproject.toml
[project]
name = "code-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <4"
dependencies = [
    "cymple>=0.12.0",
    "gitpython>=3.1.45",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langgraph>=0.6.2",
    "neo4j>=5.28.2",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "tree-sitter>=0.25.0",
    "tree-sitter-python>=0.23.6",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.0",
]
